engine: e4defrag fix error reporting
IOC_MOVE_EXT returns 0 if the requested range is beyond i_size. In that
case we currently set io_u->resid = io_u->xfer_buflen, but
io_queue_event() interprets that as a zero read and fails with EIO. We
definitely do not want this behaviour so set io_u->error explicitly to
EINVAL so we fail with that instead when we are in this scenario.